entize(ent)
The best solution for the save$ method I could suggest is that we "return" entize(ent) as ent is already modified and if there is no exception catch - that should be the response anyways. The reason why "obj = await this.entity('provider/notion/page').load$(id)" is not really recommended, after all, is it makes tests really sluggish, with the tests on the verge of timing out. Thus, the two methods come colliding. So now, there are no time outs whatsoever. The toughest test runs at 2823 ms, which is pretty good compared to the first solution.